created working general function for flash message functionality || reoccuring event undo not working!!!
