+++ app/controllers/admin/routes.py
@@ -334,17 +334,6 @@ def cancelRoute(eventId):
         
     else:
         abort(403)
-
-@admin_bp.route('/admin/undo', methods=['POST'])
-def undoEvent():
-    try:
-        eventId = session['lastDeletedEvent']
-        Event.update({Event.pendingDeletion: False}).where(Event.id == eventId).execute()
-        flash("Deletion successfully undone.", "success")
-        return " "
-    except Exception as e:
-        print('Error while canceling event:', e)
-        return "", 500
     
     
 @admin_bp.route('/event/<eventId>/delete', methods=['POST'])
